Make stack handling more robust
Attempt to fix issue #19, or at least add better logging to making
debugging it easier
* Determine "uInstructionPointer" for each stack frame, using "@$ip"
register for first frame and return address of previous frame for
others.
* Determine "uInstructionPointer" for exception from exception address.
* Perform sanity checks between exception and stack using
"uInstructionPointer", rather than the address derived from symbols.
This should be a lot more robust.
* The "uAddress" property of stack frames and exceptions is now used
exclusively informational, for addresses provided in place of symbols.
It will be None of a symbol or module(+offset) is provided. Use
"uInstructionPointer" if you need to know the exact address.