Single thread; parallel testing does not gain speed but does reduce reliability
I strongly suspect cdb is using machine-wide globals that causes it to
break in various ways when running multiple instances at the same time.
I was hoping parallel tests would speed up the tests, but this turns out
not to be the case. I strongly suspect that these machine-wide globals I
assume cdb is using contain locks that prevent it from actually running
in parallel.