Redesign interrupt-on-timeout thread
* Threads can be vital (stdin/stdout, stderr, cleanup) or helper
(interrupt-on-timeout). Vital threads run from the start to the end;
terminating any one of them terminates the entire run. Helper threads
can be created at any point and terminate at any point; the later does
not terminate the run automatically.
* The interrupt-on-timeout thread is a helper thread that is started
whenever the application is resumed, and terminates when the application
is suspended, rather than running continuously.
* `cTimeout` objects now have a description, which is supplied as the
first argument in the constrtuctor. It not used at the moment, but can
be used for debugging.
* The `oCdbWrapper` method
`fMakeSureApplicationIsInterruptedToHandleTimeouts` has been renamed to
`fAskCdbToInterruptApplication`.
* The `bOutputCanContainApplicationOutput` argument in
`fasExecuteCdbCommand` and `fasReadOutput` has been renamed to
`bApplicationWillBeRun`
* `oCdbWrapper` has a property `bApplicationIsRunnning` which is True
when the application is running in cdb and False when it is suspended.
* `oCdbWrapper` has a property `bCdbHasBeenAskedToInterruptApplication`
which is True when cBugId asked cdb to suspend the application while it
was running. It is used to identify and ignore the exception that cdb
reports when this happens.
* The `oCdbWrapper` property `oCdbLock` has been replace by
`oTimeoutAndInterruptLock`, which is a more appropriate name for its
current function.
* The `nTimeoutGranularity` setting in `dxConfig.py` was changed from
1.0 to 0.01. This should hopefully give much more accurate timeouts.