Many updarts
* drop rImportantStd*Lines: I did not use this, so it was removed to
reduce code complexity.
* drop aoInternalExceptions: cBugId uses events to report exceptions,
you should keep track of these yourselves if you need to, cBugId will no
longer do it for you.
* added "Cdb could not be terminated" Bug Report.
* Access Violations using Invalid pointers will now also be handled by
"collateral", as I found that these can be caused by `(valid heap
pointer)+(poisoned index)` so they might be controlled by an attacker.
* Windows Run Time exceptions are no longer logged in HTML reports, as I
never used that info.
* The way the cdb command line is constructed has been changed to make
it easier to read.
* Cdb is started as a `cConsoleProcess` using `mWindowsAPI`. This allows
more control than using Python subprocess and makes the code similar to
the code that starts new application processes. This allows me to use
some convenience functions that simplify the code.
* Log messages are no longer written to the HTML report throughout the
code using `fLogMessageInReport` when requested, but trigger a "Log
message" event and a single event handler is added when requested to
write them to the report.
* handling new processes has been split into handling the utility
process and application processes.
* More log messages for various events.
* Application processes are now started outside of cdb using Windows API
calls. Initial breakpoints are now ignored. The "ignored breakpoint
exceptions" for new processes has been removed. I believe this removes
any chance of mistaking an application breakpoint with a breakpoint
caused by the debugger, as there should not be any of the later anymore.