Another bunch of changes
* Main processes are now tracked by uid, and marked as such as soon as
we know their id (which may be before cdb has attached to them).
* Application debug output is now detected and "Application debug
output" events are fired so it can be processed.
* VERIFIER STOP messages are now detected through application debug
output events implemented in `cVerifierStopDetector`.
* ASan messages are no longer detected (I haven't fuzzed any ASan
compiled binaries in a while, so I haven't got a bug to re-implement it
using the application debug output event system like VERIFIER STOPs).
* "Started process" event has been renamed to "Process started".
* `cCdbWrapper.fuGetVariableId` and `cCdbWrapper.fReleaseVariableId`
have been deleted as the were not used.
* `cCdbWrapper.duAddress_by_uBreakpointId` has been deleted as it was
not used.
* `cCdbWrapper.bFatalBugDetected` has been deleted as it should no
longer be used because cBugId is now event driven.
* cdb should now be terminated with a "q" command in all but a very few
situations.
* `cBugId.fStop` (and `cCdbWrapper.fStop`) now interrupt the application
to terminate cdb using a "q" command. Note that `cCdbWrapper.fStop` does
not wait for cdb to terminate but returns immediately.
* Many `cCdbWrapper` methods have been moved to separate files.
* `cCdbWrapper.fuAddBreakpoint` has been renamed
`fuAddBreakpointForAddress` and `cCdbWrapper.fuAddBreakpointForSymbol`
has been added.
* A number of issues have been addressed that I believed to be too
unimportant to list exhaustively.
* Logging for tests has been expanded.
* excessive CPU usage checks are disabled for most checks, as they are
not expected to happen in them.