Improve Breakpoints and breakpoint exception handling
- Remove `cCdbWrapper_fuAddBreakpointForSymbol`
- Removing a breakpoints now completely removes the breakpoint in cdb,
rather than replacing it with a nop. This re-introduces a bug in cdb
that triggers breakpoint exceptions on the address where the breakpoint
was after it has been removed, but such exceptions are not ignored in
order to work around this cdb bug. This will hopefully speed up the code
for excessive CPU usage checks and fix an issue that causes random
crashes when the CPU usage worm is running.
- Create a `cException` instance in `fCdbStdInOutThread` and pass that
to `cBugReport`, rather than passing individual exception data to
`cBugReport` and have it create the instance.
- Code that works around an issue in cdb where it trigger an unwanted
initial breakpoint in a new process has been moved to a better location.
- The `cException.foCreate*` methods will not return None anymore under
any circumstance.