Updated helper threads system in attempt to fix unknown deadlock
* BugId can hang after cdb.exe stdin/out and err threads have
terminated. The cleanup thread is supposed to wait for this, clean up
and stop BugId but for some reason it did not. The new helper threads
should offer more control and logging that will help track down this
problem if it persists. It also uses locks to wait for thread
termination instead of calling threading::Thread.wait(). I suspect the
later to be the source of this issue and it may therefore have been
fixed.